#25 - Refactoring the dualis service and splitting functions into multiple files
