+++ app/src/main/java/de/fampopprol/dhbwhorb/data/dualis/network/DualisService.old.kt
@@ -1,0 +1,1389 @@
+/*
+ * SPDX-FileCopyrightText: 2024 Joinside <suitor-fall-life@duck.com>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+package de.fampopprol.dhbwhorb.data.dualis.network
+
+import android.annotation.SuppressLint
+import android.util.Log
+import de.fampopprol.dhbwhorb.data.demo.DemoDataProvider
+import de.fampopprol.dhbwhorb.data.dualis.models.DualisUrl
+import de.fampopprol.dhbwhorb.data.dualis.models.StudyGrades
+import de.fampopprol.dhbwhorb.data.dualis.models.Module
+import de.fampopprol.dhbwhorb.data.dualis.models.ExamState
+import de.fampopprol.dhbwhorb.data.dualis.models.TimetableDay
+import de.fampopprol.dhbwhorb.data.dualis.models.TimetableEvent
+import de.fampopprol.dhbwhorb.data.dualis.models.Semester
+import de.fampopprol.dhbwhorb.data.dualis.parser.StudyGradesParser
+import okhttp3.Call
+import okhttp3.Callback
+import okhttp3.FormBody
+import okhttp3.OkHttpClient
+import okhttp3.Request
+import okhttp3.Response
+import okhttp3.JavaNetCookieJar
+import java.net.CookieManager
+import java.net.CookiePolicy
+import java.io.IOException
+
+import org.jsoup.Jsoup
+
+class OldDualisService {
+
+    private val cookieManager = CookieManager(null, CookiePolicy.ACCEPT_ALL)
+    private val client = OkHttpClient.Builder().cookieJar(JavaNetCookieJar(cookieManager)).build()
+    private val _tokenRegex = Regex("ARGUMENTS=-N([0-9]{15})")
+
+    private var _authToken: String? = null
+    private var _dualisUrls: DualisUrl = DualisUrl()
+    private var _lastLoginCredentials: Pair<String, String>? = null
+    private var _isReAuthenticating = false
+    private var _isDemoMode = false
+
+    fun login(user: String, pass: String, callback: (String?) -> Unit) {
+        Log.d("DualisService", "=== STARTING LOGIN PROCESS ===")
+        Log.d("DualisService", "Username: $user")
+        Log.d("DualisService", "Password length: ${pass.length}")
+
+        // Check if this is a demo user
+        if (DemoDataProvider.isDemoUser(user) && pass == DemoDataProvider.DEMO_PASSWORD) {
+            Log.d("DualisService", "Demo user detected, enabling demo mode")
+            _isDemoMode = true
+            _lastLoginCredentials = Pair(user, pass)
+            // Simulate successful login for demo user
+            callback("Demo login successful")
+            return
+        }
+
+        // Reset demo mode for regular users
+        _isDemoMode = false
+
+        // Store credentials for potential re-authentication
+        _lastLoginCredentials = Pair(user, pass)
+
+        val formBody =
+            FormBody.Builder().add("usrname", user).add("pass", pass).add("APPNAME", "CampusNet")
+                .add("PRGNAME", "LOGINCHECK")
+                .add("ARGUMENTS", "clino,usrname,pass,menuno,menu_type,browser,platform")
+                .add("clino", "000000000000001").add("menuno", "000324").add("menu_type", "classic")
+                .add("browser", "").add("platform", "").build()
+
+        Log.d("DualisService", "Form body prepared with ${formBody.size} parameters")
+
+        val request =
+            Request.Builder().url("https://dualis.dhbw.de/scripts/mgrqispi.dll").post(formBody)
+                .build()
+
+        Log.d("DualisService", "Sending login request to: ${request.url}")
+
+        client.newCall(request).enqueue(object : Callback {
+            override fun onFailure(call: Call, e: IOException) {
+                Log.e("DualisService", "Login request failed", e)
+                callback(null)
+            }
+
+            override fun onResponse(call: Call, response: Response) {
+                val responseBody = response.body.string()
+                Log.d("DualisService", "=== LOGIN RESPONSE RECEIVED ===")
+                Log.d("DualisService", "Response code: ${response.code}")
+                Log.d("DualisService", "Response headers:")
+                response.headers.forEach { (name, value) ->
+                    Log.d("DualisService", "  $name: $value")
+                }
+                Log.d("DualisService", "Response body length: ${responseBody.length}")
+                Log.d("DualisService", "Response body preview (first 1000 chars):")
+                Log.d("DualisService", responseBody.take(1000))
+
+                if (response.isSuccessful) {
+                    val redirectUrlHeader = response.header("refresh")
+                    Log.d("DualisService", "Redirect header: $redirectUrlHeader")
+                    if (redirectUrlHeader != null) {
+                        val dualisEndpoint = "https://dualis.dhbw.de"
+                        val redirectUrlPart = if (redirectUrlHeader.contains("URL=")) {
+                            redirectUrlHeader.substring(redirectUrlHeader.indexOf("URL=") + "URL=".length)
+                        } else {
+                            redirectUrlHeader
+                        }
+                        Log.d("DualisService", "Extracted redirect URL part: $redirectUrlPart")
+                        val absoluteRedirectUrl = makeAbsoluteUrl(dualisEndpoint, redirectUrlPart)
+                        Log.d("DualisService", "Absolute redirect URL: $absoluteRedirectUrl")
+                        _updateAccessToken(absoluteRedirectUrl)
+                        followRedirects(absoluteRedirectUrl) { realMainPageContent ->
+                            if (realMainPageContent != null) {
+                                try {
+                                    parseRealMainPage(realMainPageContent)
+                                    Log.d("DualisService", "Login process completed successfully")
+                                    callback("Login successful") // Indicate success
+                                } catch (e: Exception) {
+                                    Log.e("DualisService", "Error parsing real main page", e)
+                                    callback(null)
+                                }
+                            } else {
+                                Log.e(
+                                    "DualisService",
+                                    "Real main page content is null after following redirects"
+                                )
+                                callback(null)
+                            }
+                        }
+                    } else {
+                        Log.e("DualisService", "Redirect URL is null")
+                        callback(null)
+                    }
+                } else {
+                    Log.e("DualisService", "Login failed with code: ${response.code}")
+                    callback(null)
+                }
+            }
+        })
+    }
+
+    private fun followRedirects(url: String, callback: (String?) -> Unit) {
+        val request = Request.Builder().url(url).get().build()
+
+        client.newCall(request).enqueue(object : Callback {
+            override fun onFailure(call: Call, e: IOException) {
+                Log.e("DualisService", "Follow redirects request failed", e)
+                callback(null)
+            }
+
+            override fun onResponse(call: Call, response: Response) {
+                val responseBody = response.body.string()
+                Log.d(
+                    "DualisService", "Follow Redirects Response: ${response.code} - $responseBody"
+                )
+
+                if (response.isSuccessful) {
+                    val document = Jsoup.parse(responseBody)
+                    val isRedirectPage = document.select("div#sessionId").first() != null
+                    Log.d("DualisService", "isRedirectPage check: $isRedirectPage")
+                    Log.d("DualisService", "isMainPage check: ${isMainPage(responseBody)}")
+
+                    if (isRedirectPage) {
+                        // Extract the next redirect URL
+                        var nextRedirectUrl: String? = null
+
+                        // Try to get from script first
+                        for (element in document.select("script")) {
+                            val content = element.html()
+                            if (content.contains("window.location.href")) {
+                                val regex =
+                                    Regex("window\\.location\\.href\\s*=\\s*['\"]([^'\"]+)['\"]")
+                                val match = regex.find(content)
+                                val relativeUrl = match?.groupValues?.get(1)
+                                if (relativeUrl != null) {
+                                    nextRedirectUrl = makeAbsoluteUrl(url, relativeUrl)
+                                    break
+                                }
+                            }
+                        }
+
+                        // If not found in script, try from the <a> tag
+                        if (nextRedirectUrl == null) {
+                            val anchorElement = document.select("h2 a[href]").first()
+                            val relativeUrl = anchorElement?.attr("href")
+                            if (relativeUrl != null) {
+                                nextRedirectUrl = makeAbsoluteUrl(url, relativeUrl)
+                            }
+                        }
+
+                        if (nextRedirectUrl != null) {
+                            Log.d("DualisService", "Following redirect to: $nextRedirectUrl")
+                            followRedirects(nextRedirectUrl, callback) // Recursive call
+                        } else {
+                            Log.e(
+                                "DualisService", "Could not find next redirect URL in redirect page"
+                            )
+                            callback(null)
+                        }
+                    } else if (isMainPage(responseBody)) {
+                        // This is the real main page
+                        callback(responseBody)
+                    } else {
+                        Log.e(
+                            "DualisService",
+                            "Unexpected page content, not a main page or known redirect page."
+                        )
+                        callback(null)
+                    }
+                } else {
+                    Log.e("DualisService", "Follow redirects failed with code: ${response.code}")
+                    callback(null)
+                }
+            }
+        })
+    }
+
+    private fun makeAbsoluteUrl(baseUrl: String, relativeUrl: String): String {
+        return try {
+            val base = java.net.URL(baseUrl)
+            java.net.URL(base, relativeUrl).toString()
+        } catch (e: Exception) {
+            Log.e("DualisService", "Error making absolute URL: $e")
+            ""
+        }
+    }
+
+    private fun _updateAccessToken(urlWithNewToken: String) {
+        val tokenMatch = _tokenRegex.find(urlWithNewToken)
+        if (tokenMatch != null) {
+            _authToken = tokenMatch.groupValues[1]
+            Log.d("DualisService", "Updated Auth Token: $_authToken")
+        } else {
+            Log.e("DualisService", "Auth token not found in URL: $urlWithNewToken")
+        }
+    }
+
+    private fun _fillUrlWithAuthToken(url: String): String {
+        val match = _tokenRegex.find(url)
+        return if (match != null && _authToken != null) {
+            val newUrl =
+                url.replaceRange(match.range.first, match.range.last, "ARGUMENTS=-N$_authToken")
+            Log.d("DualisService", "Filled URL with Auth Token: $newUrl")
+            newUrl
+        } else {
+            Log.w(
+                "DualisService",
+                "Could not fill URL with auth token. URL: $url, AuthToken: $_authToken"
+            )
+            url
+        }
+    }
+
+    private fun parseRealMainPage(html: String) {
+        Log.d("DualisService", "=== PARSING REAL MAIN PAGE ===")
+        Log.d("DualisService", "HTML length: ${html.length}")
+        Log.d("DualisService", "HTML content preview (first 500 chars):")
+        Log.d("DualisService", html.take(500))
+        Log.d("DualisService", "HTML content preview (last 500 chars):")
+        Log.d("DualisService", html.takeLast(500))
+
+        val document = Jsoup.parse(html)
+        Log.d("DualisService", "Document parsed successfully")
+
+        val dualisEndpoint = "https://dualis.dhbw.de"
+
+        // Log all links for debugging
+        val allLinks = document.select("a")
+        Log.d("DualisService", "Found ${allLinks.size} total links in main page:")
+        allLinks.forEachIndexed { index, link ->
+            val href = link.attr("href")
+            val text = link.text().trim()
+            Log.d("DualisService", "Link $index: text='$text', href='$href'")
+        }
+
+        // Extract the base URL pattern for COURSERESULTS from any existing link
+        // Look for links that contain COURSERESULTS pattern
+        Log.d("DualisService", "=== SEARCHING FOR COURSE RESULTS PATTERN ===")
+        var baseAuthUrl: String? = null
+
+        // Find any link that contains the auth token pattern to extract the base structure
+        allLinks.forEach { link ->
+            val href = link.attr("href")
+            if (href.contains("ARGUMENTS=-N") && href.contains("scripts/mgrqispi.dll")) {
+                baseAuthUrl = href
+                Log.d("DualisService", "Found auth token pattern in: $href")
+            }
+        }
+
+        // Construct the COURSERESULTS URL using the extracted auth token
+        if (baseAuthUrl != null && _authToken != null) {
+            // Extract the auth token and any additional arguments from the base URL
+            val baseUrl = "https://dualis.dhbw.de/scripts/mgrqispi.dll?APPNAME=CampusNet&PRGNAME=COURSERESULTS&ARGUMENTS=-N$_authToken,-N000307,"
+            _dualisUrls.studentResultsUrl = baseUrl
+            Log.d("DualisService", "Constructed student results URL: ${_dualisUrls.studentResultsUrl}")
+        } else {
+            Log.e("DualisService", "Could not construct COURSERESULTS URL - missing auth token or base URL")
+        }
+
+        // Extracting course result URL (keep existing logic for alternative)
+        Log.d("DualisService", "=== SEARCHING FOR COURSE RESULTS URL ===")
+        val courseResultElement = document.select("a:contains(Prüfungsergebnisse)").first()
+        Log.d("DualisService", "Course results element found: ${courseResultElement != null}")
+        if (courseResultElement != null) {
+            val rawHref = courseResultElement.attr("href")
+            Log.d("DualisService", "Raw course results href: '$rawHref'")
+            _dualisUrls.courseResultUrl =
+                if (rawHref.startsWith("/")) dualisEndpoint + rawHref else rawHref
+            Log.d("DualisService", "Final course results URL: '${_dualisUrls.courseResultUrl}'")
+        } else {
+            Log.e("DualisService", "No element containing 'Prüfungsergebnisse' found!")
+        }
+
+        // Extracting monthly schedule URL
+        Log.d("DualisService", "=== SEARCHING FOR SCHEDULE URL ===")
+        val monthlyScheduleElement = document.select("a:contains(diese Woche)").first()
+        Log.d("DualisService", "Monthly schedule element found: ${monthlyScheduleElement != null}")
+        if (monthlyScheduleElement != null) {
+            val rawHref = monthlyScheduleElement.attr("href")
+            Log.d("DualisService", "Raw schedule href: '$rawHref'")
+            _dualisUrls.monthlyScheduleUrl =
+                if (rawHref.startsWith("/")) dualisEndpoint + rawHref else rawHref
+            Log.d("DualisService", "Final schedule URL: '${_dualisUrls.monthlyScheduleUrl}'")
+        } else {
+            Log.e("DualisService", "No element containing 'diese Woche' found!")
+            // Try alternative searches for schedule
+            val scheduleSearches = listOf("Stundenplan", "Woche", "Schedule", "Kalender")
+            scheduleSearches.forEach { searchTerm ->
+                val scheduleElements = document.select("a:contains($searchTerm)")
+                Log.d(
+                    "DualisService",
+                    "Alternative schedule search for '$searchTerm' found ${scheduleElements.size} elements:"
+                )
+                scheduleElements.forEach { element ->
+                    Log.d(
+                        "DualisService",
+                        "  - text: '${element.text()}', href: '${element.attr("href")}'"
+                    )
+                }
+            }
+        }
+
+        // Extracting logout URL
+        Log.d("DualisService", "=== SEARCHING FOR LOGOUT URL ===")
+        val logoutElement = document.select("a:contains(Abmelden)").first()
+        Log.d("DualisService", "Logout element found: ${logoutElement != null}")
+        if (logoutElement != null) {
+            val rawHref = logoutElement.attr("href")
+            Log.d("DualisService", "Raw logout href: '$rawHref'")
+            _dualisUrls.logoutUrl =
+                if (rawHref.startsWith("/")) dualisEndpoint + rawHref else rawHref
+            Log.d("DualisService", "Final logout URL: '${_dualisUrls.logoutUrl}'")
+        } else {
+            Log.e("DualisService", "No element containing 'Abmelden' found!")
+        }
+
+        Log.d("DualisService", "=== FINAL PARSED DUALIS URLS ===")
+        Log.d("DualisService", "Student Results URL: ${_dualisUrls.studentResultsUrl}")
+        Log.d("DualisService", "Course Results URL: ${_dualisUrls.courseResultUrl}")
+        Log.d("DualisService", "Monthly Schedule URL: ${_dualisUrls.monthlyScheduleUrl}")
+        Log.d("DualisService", "Logout URL: ${_dualisUrls.logoutUrl}")
+        Log.d("DualisService", "=== END MAIN PAGE PARSING ===")
+    }
+
+    @SuppressLint("DefaultLocale")
+    fun getMonthlySchedule(year: Int, month: Int, callback: (List<TimetableDay>?) -> Unit) {
+        // Return demo data if in demo mode
+        if (_isDemoMode) {
+            Log.d(
+                "DualisService", "Demo mode: returning demo timetable data for month $month/$year"
+            )
+            // For monthly view, generate demo data for the first week of the month
+            val firstDayOfMonth = java.time.LocalDate.of(year, month, 1)
+            val firstMonday =
+                firstDayOfMonth.with(java.time.temporal.TemporalAdjusters.previousOrSame(java.time.DayOfWeek.MONDAY))
+            val demoData = DemoDataProvider.getDemoTimetableForWeek(firstMonday)
+            callback(demoData)
+            return
+        }
+
+        if (_dualisUrls.monthlyScheduleUrl == null || _authToken == null) {
+            Log.e(
+                "DualisService",
+                "Monthly schedule URL or Auth Token is null. Cannot fetch timetable."
+            )
+            callback(null)
+            return
+        }
+
+        val baseUrl = _dualisUrls.monthlyScheduleUrl!!
+        val authToken = _authToken!!
+
+        // Extract existing arguments
+        val argumentsRegex = Regex("ARGUMENTS=([^&]+)")
+        val existingArgumentsMatch = argumentsRegex.find(baseUrl)
+        val existingArguments = existingArgumentsMatch?.groupValues?.get(1) ?: ""
+
+        // Format the date as dd.MM.yyyy for the first day of the month
+        val formattedDate = String.format("%02d.%02d.%d", 1, month, year)
+
+        // Replace the first -A with the formatted date
+        val updatedArguments = existingArguments.replaceFirst("-A", "-A$formattedDate")
+
+        // Reconstruct the URL
+        val url = baseUrl.replace(existingArguments, updatedArguments).replace(
+            "ARGUMENTS=-N$authToken", "ARGUMENTS=-N$authToken"
+        ) // Ensure auth token is correct
+        Log.d("DualisService", "Constructed Monthly Schedule URL: $url")
+
+        val request = Request.Builder().url(url).get().build()
+
+        client.newCall(request).enqueue(object : Callback {
+            override fun onFailure(call: Call, e: IOException) {
+                Log.e("DualisService", "Get monthly schedule request failed", e)
+                callback(null)
+            }
+
+            override fun onResponse(call: Call, response: Response) {
+                val responseBody = response.body.string()
+                Log.d(
+                    "DualisService", "Monthly Schedule Response: ${response.code} - $responseBody"
+                )
+
+                if (response.isSuccessful) {
+                    try {
+                        val timetableDays = parseMonthlySchedule(responseBody)
+                        callback(timetableDays)
+                    } catch (e: Exception) {
+                        Log.e("DualisService", "Error parsing monthly schedule", e)
+                        callback(null)
+                    }
+                } else {
+                    Log.e(
+                        "DualisService", "Get monthly schedule failed with code: ${response.code}"
+                    )
+                    callback(null)
+                }
+            }
+        })
+    }
+
+    fun getWeeklySchedule(
+        targetDate: java.time.LocalDate, callback: (List<TimetableDay>?) -> Unit
+    ) {
+        // Return demo data if in demo mode
+        if (_isDemoMode) {
+            Log.d(
+                "DualisService",
+                "Demo mode: returning demo timetable data for week starting $targetDate"
+            )
+            val demoData = DemoDataProvider.getDemoTimetableForWeek(targetDate)
+            callback(demoData)
+            return
+        }
+
+        getWeeklyScheduleWithRetry(targetDate, callback, retryCount = 0)
+    }
+
+    private fun getWeeklyScheduleWithRetry(
+        targetDate: java.time.LocalDate, callback: (List<TimetableDay>?) -> Unit, retryCount: Int
+    ) {
+        if (_dualisUrls.monthlyScheduleUrl == null || _authToken == null) {
+            Log.e(
+                "DualisService",
+                "Monthly schedule URL or Auth Token is null. Cannot fetch weekly timetable."
+            )
+            callback(null)
+            return
+        }
+
+        val baseUrl = _dualisUrls.monthlyScheduleUrl!!
+        val authToken = _authToken!!
+
+        // Extract existing arguments
+        val argumentsRegex = Regex("ARGUMENTS=([^&]+)")
+        val existingArgumentsMatch = argumentsRegex.find(baseUrl)
+        val existingArguments = existingArgumentsMatch?.groupValues?.get(1) ?: ""
+
+        // Format the target date as dd.MM.yyyy
+        val dateFormatter = java.time.format.DateTimeFormatter.ofPattern("dd.MM.yyyy")
+        val formattedDate = targetDate.format(dateFormatter)
+
+        // Replace the first -A with the formatted target date
+        val updatedArguments = existingArguments.replaceFirst("-A", "-A$formattedDate")
+
+        // Reconstruct the URL
+        val url = baseUrl.replace(existingArguments, updatedArguments).replace(
+            "ARGUMENTS=-N$authToken", "ARGUMENTS=-N$authToken"
+        ) // Ensure auth token is correct
+        Log.d("DualisService", "Constructed Weekly Schedule URL for $targetDate: $url")
+
+        val request = Request.Builder().url(url).get().build()
+
+        client.newCall(request).enqueue(object : Callback {
+            override fun onFailure(call: Call, e: IOException) {
+                Log.e("DualisService", "Get weekly schedule request failed", e)
+                callback(null)
+            }
+
+            override fun onResponse(call: Call, response: Response) {
+                val responseBody = response.body.string()
+                Log.d("DualisService", "Weekly Schedule Response: ${response.code}")
+
+                if (response.isSuccessful) {
+                    // Check if the response indicates an invalid token
+                    if (isTokenInvalidResponse(responseBody)) {
+                        Log.w(
+                            "DualisService",
+                            "Token appears to be invalid, attempting re-authentication"
+                        )
+                        if (retryCount < 1) { // Only retry once
+                            reAuthenticateIfNeeded { success ->
+                                if (success) {
+                                    Log.d(
+                                        "DualisService",
+                                        "Re-authentication successful, retrying weekly schedule fetch"
+                                    )
+                                    getWeeklyScheduleWithRetry(targetDate, callback, retryCount + 1)
+                                } else {
+                                    Log.e("DualisService", "Re-authentication failed")
+                                    callback(null)
+                                }
+                            }
+                        } else {
+                            Log.e("DualisService", "Already retried once, giving up")
+                            callback(null)
+                        }
+                        return
+                    }
+
+                    try {
+                        val timetableDays = parseMonthlySchedule(responseBody)
+                        Log.d(
+                            "DualisService",
+                            "Parsed weekly schedule for $targetDate: ${timetableDays.size} days"
+                        )
+
+                        // Enhance timetable with detailed information from individual event pages
+                        enhanceTimetableWithDetails(timetableDays) { enhancedTimetableDays ->
+                            if (enhancedTimetableDays != null) {
+                                Log.d(
+                                    "DualisService",
+                                    "Enhanced weekly schedule for $targetDate with detailed information"
+                                )
+                                callback(enhancedTimetableDays)
+                            } else {
+                                Log.w(
+                                    "DualisService",
+                                    "Failed to enhance timetable with details, returning basic timetable"
+                                )
+                                callback(timetableDays)
+                            }
+                        }
+                    } catch (e: Exception) {
+                        Log.e("DualisService", "Error parsing weekly schedule", e)
+                        callback(null)
+                    }
+                } else {
+                    Log.e("DualisService", "Get weekly schedule failed with code: ${response.code}")
+                    callback(null)
+                }
+            }
+        })
+    }
+
+    /**
+     * Fetches available semesters from Dualis
+     */
+    fun getAvailableSemesters(callback: (List<Semester>?) -> Unit) {
+        Log.d("DualisService", "=== FETCHING AVAILABLE SEMESTERS ===")
+
+        // Return demo data if in demo mode
+        if (_isDemoMode) {
+            Log.d("DualisService", "Demo mode: returning demo semesters")
+            callback(Semester.getDefaultSemesters())
+            return
+        }
+
+        if (_authToken == null) {
+            Log.e("DualisService", "Auth Token is null. Authentication required.")
+            callback(null)
+            return
+        }
+
+        // Use the base URL for course results to get the semester selection page
+        val baseUrl = "https://dualis.dhbw.de/scripts/mgrqispi.dll?APPNAME=CampusNet&PRGNAME=COURSERESULTS&ARGUMENTS=-N$_authToken,-N000307,"
+
+        val request = Request.Builder().url(baseUrl).get().build()
+
+        Log.d("DualisService", "Fetching semesters from: $baseUrl")
+
+        client.newCall(request).enqueue(object : Callback {
+            override fun onFailure(call: Call, e: IOException) {
+                Log.e("DualisService", "Get available semesters request failed", e)
+                callback(null)
+            }
+
+            override fun onResponse(call: Call, response: Response) {
+                val responseBody = response.body.string()
+                Log.d("DualisService", "Semesters Response: ${response.code}")
+
+                if (response.isSuccessful) {
+                    try {
+                        val semesters = parseSemestersFromHtml(responseBody)
+                        Log.d("DualisService", "Parsed ${semesters.size} semesters")
+                        callback(semesters)
+                    } catch (e: Exception) {
+                        Log.e("DualisService", "Error parsing semesters", e)
+                        // Fallback to default semesters if parsing fails
+                        callback(Semester.getDefaultSemesters())
+                    }
+                } else {
+                    Log.e("DualisService", "Get semesters failed with code: ${response.code}")
+                    // Fallback to default semesters if request fails
+                    callback(Semester.getDefaultSemesters())
+                }
+            }
+        })
+    }
+
+    /**
+     * Parse semesters from the HTML response containing the semester dropdown
+     */
+    private fun parseSemestersFromHtml(html: String): List<Semester> {
+        Log.d("DualisService", "=== PARSING SEMESTERS FROM HTML ===")
+
+        val document = Jsoup.parse(html)
+        val semesters = mutableListOf<Semester>()
+
+        // Look for semester select dropdown
+        val semesterSelect = document.select("select#semester").first()
+
+        if (semesterSelect != null) {
+            Log.d("DualisService", "Found semester select dropdown")
+
+            val options = semesterSelect.select("option")
+            Log.d("DualisService", "Found ${options.size} semester options")
+
+            options.forEach { option ->
+                val value = option.attr("value")
+                val displayName = option.text().trim()
+                val isSelected = option.hasAttr("selected")
+
+                if (value.isNotEmpty() && displayName.isNotEmpty()) {
+                    semesters.add(Semester(value, displayName, isSelected))
+                    Log.d("DualisService", "Added semester: $displayName (value: $value, selected: $isSelected)")
+                }
+            }
+        } else {
+            Log.w("DualisService", "No semester select dropdown found, using default semesters")
+            return Semester.getDefaultSemesters()
+        }
+
+        // If no semesters were found, use defaults
+        if (semesters.isEmpty()) {
+            Log.w("DualisService", "No semesters parsed, using default semesters")
+            return Semester.getDefaultSemesters()
+        }
+
+        Log.d("DualisService", "Successfully parsed ${semesters.size} semesters")
+        return semesters
+    }
+
+    /**
+     * Fetches the study grades for a specific semester
+     */
+    fun getStudyGradesForSemester(semester: Semester, callback: (StudyGrades?) -> Unit) {
+        val semesterArgument = Semester.formatSemesterArgument(semester.value)
+        Log.d("DualisService", "Fetching grades for semester: ${semester.displayName} with argument: $semesterArgument")
+        getStudyGrades(semesterArgument, callback)
+    }
+
+    /**
+     * Fetches the study grades (GPA and credits information) from Dualis for a specific semester
+     */
+    fun getStudyGrades(semesterArgument: String = "", callback: (StudyGrades?) -> Unit) {
+        Log.d("DualisService", "=== STARTING STUDY GRADES FETCH ===")
+        Log.d("DualisService", "Semester argument: $semesterArgument")
+
+        // Return demo data if in demo mode
+        if (_isDemoMode) {
+            Log.d("DualisService", "Demo mode: returning demo study grades data")
+            callback(
+                StudyGrades(
+                    gpaTotal = 1.7,
+                    gpaMainModules = 1.6,
+                    creditsTotal = 210.0,
+                    creditsGained = 180.0,
+                    modules = listOf(
+                        Module(
+                            id = "T4_1000",
+                            name = "Praxisprojekt I",
+                            credits = "20.0",
+                            grade = "1.3",
+                            state = ExamState.PASSED
+                        ),
+                        Module(
+                            id = "T4INF1003",
+                            name = "Theoretische Informatik II",
+                            credits = "5.0",
+                            grade = "noch nicht gesetzt",
+                            state = ExamState.PENDING
+                        )
+                    ),
+                    semester = if (semesterArgument.isEmpty()) "current" else "previous"
+                )
+            )
+            return
+        }
+
+        // Check authentication status
+        Log.d("DualisService", "Checking authentication status...")
+        Log.d("DualisService", "Auth Token present: ${_authToken != null}")
+        Log.d("DualisService", "Auth Token value: $_authToken")
+        Log.d("DualisService", "isAuthenticated(): ${isAuthenticated()}")
+
+        if (_authToken == null) {
+            Log.e("DualisService", "Auth Token is null. Authentication required.")
+            callback(null)
+            return
+        }
+
+        Log.d("DualisService", "Checking student results URL...")
+        Log.d("DualisService", "Student results URL present: ${_dualisUrls.studentResultsUrl != null}")
+        Log.d("DualisService", "Student results URL value: ${_dualisUrls.studentResultsUrl}")
+
+        if (_dualisUrls.studentResultsUrl == null) {
+            Log.e("DualisService", "Student results URL is null. Main page parsing may have failed.")
+            Log.e("DualisService", "Available URLs in _dualisUrls:")
+            Log.e("DualisService", "  - studentResultsUrl: ${_dualisUrls.studentResultsUrl}")
+            Log.e("DualisService", "  - courseResultUrl: ${_dualisUrls.courseResultUrl}")
+            Log.e("DualisService", "  - monthlyScheduleUrl: ${_dualisUrls.monthlyScheduleUrl}")
+            Log.e("DualisService", "  - logoutUrl: ${_dualisUrls.logoutUrl}")
+            callback(null)
+            return
+        }
+
+        // Construct URL with semester-specific argument
+        // Base format: https://dualis.dhbw.de/scripts/mgrqispi.dll?APPNAME=CampusNet&PRGNAME=COURSERESULTS&ARGUMENTS=-N{authToken},-N000307{semesterArgument}
+        val baseUrl = "https://dualis.dhbw.de/scripts/mgrqispi.dll?APPNAME=CampusNet&PRGNAME=COURSERESULTS&ARGUMENTS=-N$_authToken,-N000307$semesterArgument"
+
+        Log.d("DualisService", "=== URL PREPARATION COMPLETE ===")
+        Log.d("DualisService", "Auth Token: $_authToken")
+        Log.d("DualisService", "Semester-specific URL: $baseUrl")
+
+        val request = Request.Builder().url(baseUrl).get().build()
+
+        Log.d("DualisService", "=== SENDING HTTP REQUEST ===")
+        Log.d("DualisService", "Request URL: ${request.url}")
+        Log.d("DualisService", "Request method: ${request.method}")
+
+        client.newCall(request).enqueue(object : Callback {
+            override fun onFailure(call: Call, e: IOException) {
+                Log.e("DualisService", "=== HTTP REQUEST FAILED ===")
+                Log.e("DualisService", "Request URL: ${call.request().url}")
+                Log.e("DualisService", "Error type: ${e.javaClass.simpleName}")
+                Log.e("DualisService", "Error message: ${e.message}")
+                Log.e("DualisService", "Get study grades request failed", e)
+                callback(null)
+            }
+
+            override fun onResponse(call: Call, response: Response) {
+                val responseBody = response.body.string()
+                Log.d("DualisService", "=== HTTP RESPONSE RECEIVED ===")
+                Log.d("DualisService", "Response code: ${response.code}")
+                Log.d("DualisService", "Response message: ${response.message}")
+                Log.d("DualisService", "Response headers:")
+                response.headers.forEach { (name, value) ->
+                    Log.d("DualisService", "  $name: $value")
+                }
+                Log.d("DualisService", "Response body length: ${responseBody.length}")
+                Log.d("DualisService", "Response body preview (first 1000 chars):")
+                Log.d("DualisService", responseBody.take(1000))
+                Log.d("DualisService", "Response body preview (last 500 chars):")
+                Log.d("DualisService", responseBody.takeLast(500))
+
+                if (response.isSuccessful) {
+                    Log.d("DualisService", "=== RESPONSE SUCCESSFUL - PROCESSING ===")
+
+                    // Check if the response indicates an invalid token
+                    val isTokenInvalid = isTokenInvalidResponse(responseBody)
+                    Log.d("DualisService", "Token invalid check result: $isTokenInvalid")
+
+                    if (isTokenInvalid) {
+                        Log.w("DualisService", "Token appears to be invalid when fetching grades, attempting re-authentication")
+                        reAuthenticateIfNeeded { success ->
+                            if (success) {
+                                Log.d("DualisService", "Re-authentication successful, retrying grades fetch")
+                                getStudyGrades(semesterArgument, callback) // Retry after re-authentication
+                            } else {
+                                Log.e("DualisService", "Re-authentication failed")
+                                callback(null)
+                            }
+                        }
+                        return
+                    }
+
+                    try {
+                        Log.d("DualisService", "=== STARTING HTML PARSING ===")
+                        val parser = StudyGradesParser()
+                        val studyGrades = parser.extractStudyGrades(responseBody, semesterArgument)
+                        Log.d("DualisService", "=== PARSING COMPLETE ===")
+                        Log.d("DualisService", "Parsed study grades: $studyGrades")
+
+                        if (studyGrades != null) {
+                            Log.d("DualisService", "Successfully parsed grades:")
+                            Log.d("DualisService", "  - GPA Total: ${studyGrades.gpaTotal}")
+                            Log.d("DualisService", "  - GPA Main Modules: ${studyGrades.gpaMainModules}")
+                            Log.d("DualisService", "  - Credits Total: ${studyGrades.creditsTotal}")
+                            Log.d("DualisService", "  - Credits Gained: ${studyGrades.creditsGained}")
+                            Log.d("DualisService", "  - Semester: ${studyGrades.semester}")
+                        } else {
+                            Log.e("DualisService", "Parser returned null - parsing failed")
+                        }
+
+                        callback(studyGrades)
+                    } catch (e: Exception) {
+                        Log.e("DualisService", "=== PARSING ERROR ===")
+                        Log.e("DualisService", "Error type: ${e.javaClass.simpleName}")
+                        Log.e("DualisService", "Error message: ${e.message}")
+                        Log.e("DualisService", "Error parsing study grades", e)
+                        callback(null)
+                    }
+                } else {
+                    Log.e("DualisService", "=== HTTP RESPONSE ERROR ===")
+                    Log.e("DualisService", "Response code: ${response.code}")
+                    Log.e("DualisService", "Response message: ${response.message}")
+                    Log.e("DualisService", "Get study grades failed with code: ${response.code}")
+                    callback(null)
+                }
+            }
+        })
+    }
+
+    private fun parseMonthlySchedule(html: String): List<TimetableDay> {
+        val document = Jsoup.parse(html)
+        val dateFormatter = java.time.format.DateTimeFormatter.ofPattern("dd.MM.yyyy")
+
+        val table = document.select("table.nb").first() ?: return emptyList()
+        val caption = table.select("caption").first()?.text()
+        Log.d("DualisService", "Table caption: $caption")
+
+        val dateRangeRegex = Regex("Stundenplan vom (\\d{2}\\.\\d{2}\\.) bis (\\d{2}\\.\\d{2}\\.)")
+        val matchResult = caption?.let { dateRangeRegex.find(it) }
+
+        val startDateString = matchResult?.groupValues?.get(1)
+        val endDateString = matchResult?.groupValues?.get(2)
+        Log.d(
+            "DualisService", "Start date string: $startDateString, End date string: $endDateString"
+        )
+
+        val currentYear = java.time.LocalDate.now().year
+
+        val startLocalDate = requireNotNull(startDateString?.let {
+            java.time.LocalDate.parse(
+                it + currentYear, dateFormatter
+            )
+        }) {
+            "Could not parse start date from caption: $caption"
+        }
+        val endLocalDate = requireNotNull(endDateString?.let {
+            java.time.LocalDate.parse(
+                it + currentYear, dateFormatter
+            )
+        }) {
+            "Could not parse end date from caption: $caption"
+        }
+
+        // Find the header row with weekday columns
+        val headerRow = table.select("tr.tbsubhead").first() ?: return emptyList()
+        val dayToDateMap = mutableMapOf<String, java.time.LocalDate>()
+
+        // Parse dates from table headers directly - look for th.weekday elements with links
+        headerRow.select("th.weekday").forEach { dayHeaderElement ->
+            val link = dayHeaderElement.select("a").first()
+            val headerText = link?.text()?.trim() ?: dayHeaderElement.text().trim()
+            Log.d("DualisService", "Processing header: '$headerText'")
+
+            // Extract day abbreviation and date from header text like "Mo 30.06."
+            val headerPattern = Regex("(\\w+)\\s+(\\d{2}\\.\\d{2})\\.")
+            val headerMatch = headerPattern.find(headerText)
+
+            if (headerMatch != null) {
+                val dayAbbreviation = headerMatch.groupValues[1]
+                val dateString = headerMatch.groupValues[2] + ".$currentYear"
+
+                val fullDayName = when (dayAbbreviation) {
+                    "Mo" -> "Montag"
+                    "Di" -> "Dienstag"
+                    "Mi" -> "Mittwoch"
+                    "Do" -> "Donnerstag"
+                    "Fr" -> "Freitag"
+                    "Sa" -> "Samstag"
+                    "So" -> "Sonntag"
+                    else -> {
+                        Log.w("DualisService", "Unknown day abbreviation: $dayAbbreviation")
+                        ""
+                    }
+                }
+
+                if (fullDayName.isNotEmpty()) {
+                    try {
+                        val parsedDate = java.time.LocalDate.parse(dateString, dateFormatter)
+                        dayToDateMap[fullDayName] = parsedDate
+                        Log.d("DualisService", "Mapped $fullDayName to $parsedDate")
+                    } catch (e: Exception) {
+                        Log.e("DualisService", "Error parsing date: $dateString", e)
+                    }
+                }
+            } else {
+                Log.w("DualisService", "Could not parse header: '$headerText'")
+            }
+        }
+
+        // If no headers were found with the standard approach, try extracting directly from the range
+        if (dayToDateMap.isEmpty()) {
+            Log.w("DualisService", "No dates found in headers, trying to map from date range")
+
+            // Create date mapping based on the date range from caption
+            var currentDate = startLocalDate
+            val weekDays = listOf(
+                "Montag", "Dienstag", "Mittwoch", "Donnerstag", "Freitag", "Samstag", "Sonntag"
+            )
+
+            while (!currentDate.isAfter(endLocalDate)) {
+                val dayOfWeek = currentDate.dayOfWeek.value // 1 = Monday, 7 = Sunday
+                val dayName = weekDays[dayOfWeek - 1]
+                dayToDateMap[dayName] = currentDate
+                Log.d("DualisService", "Fallback mapped $dayName to $currentDate")
+                currentDate = currentDate.plusDays(1)
+            }
+        }
+
+        Log.d("DualisService", "Day To Date Map: $dayToDateMap")
+
+        val eventsByFullDate = mutableMapOf<java.time.LocalDate, MutableList<TimetableEvent>>()
+        var currentDay = startLocalDate
+        while (!currentDay.isAfter(endLocalDate)) {
+            eventsByFullDate[currentDay] = mutableListOf()
+            currentDay = currentDay.plusDays(1)
+        }
+
+        val allAppointmentCells = document.select("td.appointment")
+        Log.d("DualisService", "Found ${allAppointmentCells.size} appointment cells")
+
+        // Parse events - for now just create basic events without detailed fetching
+        // (We'll implement asynchronous detail fetching in the calling methods)
+        for (cell in allAppointmentCells) {
+            val cellHtml = cell.html()
+            Log.d("DualisService", "Processing appointment cell HTML: $cellHtml")
+
+            // Extract event detail URL from links in the cell
+            val eventLink = cell.select("a").first()
+            var eventDetailUrl: String? = null
+
+            if (eventLink != null) {
+                val href = eventLink.attr("href")
+                if (href.isNotEmpty()) {
+                    eventDetailUrl = if (href.startsWith("/")) {
+                        "https://dualis.dhbw.de$href"
+                    } else if (href.startsWith("scripts/")) {
+                        "https://dualis.dhbw.de/$href"
+                    } else if (!href.startsWith("http")) {
+                        "https://dualis.dhbw.de/scripts/$href"
+                    } else {
+                        href
+                    }
+                    Log.d("DualisService", "Found event detail URL: $eventDetailUrl")
+                }
+            }
+
+            // Extract title - get text content excluding timePeriod span
+            val clonedCell = cell.clone()
+            clonedCell.select("span.timePeriod").remove()
+            clonedCell.select("br").remove()
+            // Remove HTML comments and clean up
+            var title = clonedCell.text().trim()
+
+            // Clean up title by removing trailing ">" and any HTML artifacts
+            title = title.replace(Regex(">\\s*$"), "").trim()
+
+            // Skip if title is empty
+            if (title.isEmpty()) {
+                Log.d("DualisService", "Skipping cell with empty title")
+                continue
+            }
+
+            // Extract time and room information
+            val timePeriodSpans = cell.select("span.timePeriod")
+            var timePeriodText = ""
+
+            // Combine text from all timePeriod spans
+            for (span in timePeriodSpans) {
+                val spanText = span.text().trim()
+                if (spanText.isNotEmpty()) {
+                    if (timePeriodText.isNotEmpty()) {
+                        timePeriodText += " $spanText"
+                    } else {
+                        timePeriodText = spanText
+                    }
+                }
+            }
+
+            Log.d("DualisService", "Time period text: '$timePeriodText'")
+
+            // Parse time period - it might be in format "08:15 - 12:30 HOR-120" or similar
+            val timeRoomParts = timePeriodText.split("\\s+".toRegex()).filter { it.isNotBlank() }
+
+            var startTime = ""
+            var endTime = ""
+            var room = ""
+
+            if (timeRoomParts.size >= 3) {
+                startTime = timeRoomParts[0]
+                // Skip the "-" separator
+                endTime = timeRoomParts[2]
+                // Room might be in the remaining parts
+                if (timeRoomParts.size > 3) {
+                    val rawRoom = timeRoomParts.drop(3).joinToString(" ")
+                    room = parseRooms(rawRoom)
+                }
+            }
+
+            val lecturer = "" // Will be filled from detailed information later
+
+            // Get the day from the abbr attribute
+            val abbrAttribute = cell.attr("abbr")
+            val dayOfWeekInGerman = abbrAttribute.split(" ")[0]
+
+            val eventDate = dayToDateMap[dayOfWeekInGerman]
+
+            Log.d("DualisService", "Processing cell:")
+            Log.d("DualisService", "  Title: '$title'")
+            Log.d("DualisService", "  Time Period Text: '$timePeriodText'")
+            Log.d("DualisService", "  Time Room Parts: $timeRoomParts")
+            Log.d(
+                "DualisService", "  Start Time: '$startTime', End Time: '$endTime', Room: '$room'"
+            )
+            Log.d(
+                "DualisService",
+                "  Abbr Attribute: '$abbrAttribute', Day in German: '$dayOfWeekInGerman'"
+            )
+            Log.d("DualisService", "  Event Date: $eventDate")
+            Log.d("DualisService", "  Event Detail URL: $eventDetailUrl")
+
+            if (eventDate != null && title.isNotEmpty()) {
+                // Create event object with basic information
+                val event = TimetableEvent(
+                    title = title,
+                    startTime = startTime,
+                    endTime = endTime,
+                    room = room,
+                    lecturer = lecturer,
+                    detailUrl = eventDetailUrl
+                )
+
+                eventsByFullDate[eventDate]?.add(event)
+                Log.d("DualisService", "Added event to date $eventDate: $title")
+            } else {
+                Log.w("DualisService", "Skipping event - eventDate: $eventDate, title: '$title'")
+            }
+        }
+
+        val sortedTimetableDays = eventsByFullDate.entries.sortedBy { it.key }.map { entry ->
+            Log.d(
+                "DualisService",
+                "Creating TimetableDay for ${dateFormatter.format(entry.key)} with ${entry.value.size} events"
+            )
+            TimetableDay(dateFormatter.format(entry.key), entry.value)
+        }
+
+        Log.d("DualisService", "Parsed ${sortedTimetableDays.size} timetable days")
+        sortedTimetableDays.forEach { day ->
+            Log.d("DualisService", "Day ${day.date}: ${day.events.size} events")
+            day.events.forEach { event ->
+                Log.d(
+                    "DualisService",
+                    "  Event: ${event.title} (${event.startTime} - ${event.endTime}) in ${event.room}"
+                )
+            }
+        }
+
+        return sortedTimetableDays
+    }
+
+    /**
+     * Enhances timetable events with detailed information from individual event pages
+     */
+    fun enhanceTimetableWithDetails(
+        timetableDays: List<TimetableDay>,
+        callback: (List<TimetableDay>?) -> Unit
+    ) {
+        if (_isDemoMode || timetableDays.isEmpty()) {
+            // Return original data for demo mode or empty lists
+            callback(timetableDays)
+            return
+        }
+
+        Log.d("DualisService", "=== ENHANCING TIMETABLE WITH DETAILED INFORMATION ===")
+
+        // Collect all events that have detail URLs
+        val eventsToEnhance = mutableListOf<Triple<TimetableEvent, TimetableDay, Int>>()
+
+        timetableDays.forEach { day ->
+            day.events.forEachIndexed { index, event ->
+                if (event.detailUrl != null) {
+                    eventsToEnhance.add(Triple(event, day, index))
+                }
+            }
+        }
+
+        if (eventsToEnhance.isEmpty()) {
+            Log.d("DualisService", "No events with detail URLs found, returning original data")
+            callback(timetableDays)
+            return
+        }
+
+        Log.d("DualisService", "Found ${eventsToEnhance.size} events to enhance with detailed information")
+
+        // Create enhanced versions of the timetable days with mutable event lists
+        val enhancedDays = timetableDays.map { day ->
+            TimetableDay(day.date, day.events.toMutableList())
+        }
+
+        var completedRequests = 0
+        val totalRequests = eventsToEnhance.size
+
+        // Fetch details for each event
+        eventsToEnhance.forEach { (event, originalDay, eventIndex) ->
+            fetchEventDetails(event.detailUrl!!) { eventDetails ->
+                synchronized(enhancedDays) {
+                    completedRequests++
+
+                    // Find the corresponding day in enhanced days
+                    val enhancedDay = enhancedDays.find { it.date == originalDay.date }
+
+                    if (enhancedDay != null && eventIndex < enhancedDay.events.size) {
+                        val enhancedEvent = if (eventDetails != null) {
+                            event.copy(
+                                fullTitle = eventDetails.fullTitle,
+                                courseCode = eventDetails.courseCode,
+                                lecturer = eventDetails.lecturer,
+                                room = eventDetails.room.ifEmpty { event.room }
+                            )
+                        } else {
+                            // Keep original event if details couldn't be fetched
+                            event
+                        }
+
+                        // Replace the event in the enhanced day (now using MutableList)
+                        (enhancedDay.events as MutableList)[eventIndex] = enhancedEvent
+
+                        Log.d("DualisService", "Enhanced event ${event.title} -> ${enhancedEvent.fullTitle ?: event.title}")
+                        if (eventDetails != null) {
+                            Log.d("DualisService", "  Course code: ${eventDetails.courseCode}")
+                            Log.d("DualisService", "  Lecturer: ${eventDetails.lecturer}")
+                            Log.d("DualisService", "  Room: ${eventDetails.room}")
+                        }
+                    }
+
+                    // Check if all requests are completed
+                    if (completedRequests >= totalRequests) {
+                        Log.d("DualisService", "All event details fetched, calling callback with enhanced data")
+                        callback(enhancedDays)
+                    }
+
+                    Log.d("DualisService", "Enhancement progress: $completedRequests/$totalRequests")
+                }
+            }
+        }
+    }
+
+    /**
+     * Fetches detailed information for pending events
+     */
+    private fun fetchEventDetailsForPendingEvents(
+        pendingEvents: List<Pair<TimetableEvent, java.time.LocalDate>>,
+        eventsByFullDate: MutableMap<java.time.LocalDate, MutableList<TimetableEvent>>
+    ) {
+        var completedRequests = 0
+        val totalRequests = pendingEvents.size
+
+        for ((event, eventDate) in pendingEvents) {
+            if (event.detailUrl != null) {
+                fetchEventDetails(event.detailUrl) { eventDetails ->
+                    completedRequests++
+
+                    // Create enhanced event with detailed information
+                    val enhancedEvent = if (eventDetails != null) {
+                        event.copy(
+                            fullTitle = eventDetails.fullTitle,
+                            courseCode = eventDetails.courseCode,
+                            lecturer = eventDetails.lecturer,
+                            room = if (eventDetails.room.isNotEmpty()) eventDetails.room else event.room
+                        )
+                    } else {
+                        // Use basic event if details couldn't be fetched
+                        event
+                    }
+
+                    // Add the enhanced event to the appropriate date
+                    synchronized(eventsByFullDate) {
+                        eventsByFullDate[eventDate]?.add(enhancedEvent)
+                        Log.d("DualisService", "Added enhanced event to date $eventDate: ${enhancedEvent.title}")
+                        if (eventDetails != null) {
+                            Log.d("DualisService", "  Full title: ${eventDetails.fullTitle}")
+                            Log.d("DualisService", "  Course code: ${eventDetails.courseCode}")
+                            Log.d("DualisService", "  Lecturer: ${eventDetails.lecturer}")
+                        }
+                    }
+
+                    // Log progress
+                    Log.d("DualisService", "Event details fetch progress: $completedRequests/$totalRequests")
+                }
+            } else {
+                // No detail URL, add basic event
+                eventsByFullDate[eventDate]?.add(event)
+                completedRequests++
+                Log.d("DualisService", "Added basic event (no detail URL) to date $eventDate: ${event.title}")
+            }
+        }
+    }
+
+    /**
+     * Fetches detailed information for a specific event from its detail page
+     */
+    private fun fetchEventDetails(eventUrl: String, callback: (EventDetails?) -> Unit) {
+        if (_isDemoMode) {
+            // Return null for demo mode - we'll use basic info only
+            callback(null)
+            return
+        }
+
+        val request = Request.Builder().url(eventUrl).get().build()
+
+        client.newCall(request).enqueue(object : Callback {
+            override fun onFailure(call: Call, e: IOException) {
+                Log.e("DualisService", "Failed to fetch event details from: $eventUrl", e)
+                callback(null)
+            }
+
+            override fun onResponse(call: Call, response: Response) {
+                val responseBody = response.body.string()
+                Log.d("DualisService", "Event details response code: ${response.code}")
+
+                if (response.isSuccessful) {
+                    try {
+                        val eventDetails = parseEventDetails(responseBody)
+                        callback(eventDetails)
+                    } catch (e: Exception) {
+                        Log.e("DualisService", "Error parsing event details", e)
+                        callback(null)
+                    }
+                } else {
+                    Log.e("DualisService", "Failed to fetch event details with code: ${response.code}")
+                    callback(null)
+                }
+            }
+        })
+    }
+
+    /**
+     * Parses event details from the HTML of an individual event page
+     */
+    private fun parseEventDetails(html: String): EventDetails? {
+        val document = Jsoup.parse(html)
+
+        // Extract the full title from h1 tag
+        val h1Element = document.select("h1").first()
+        val fullTitle = h1Element?.text()?.trim()
+
+        if (fullTitle.isNullOrEmpty()) {
+            Log.w("DualisService", "Could not extract full title from event details page")
+            return null
+        }
+
+        // Extract course code from the full title (e.g., "T4INF1003.1" from "T4INF1003.1 Algorithmen und Komplexität HOR-TINF2024")
+        val courseCodePattern = Regex("^([A-Z0-9.]+)")
+        val courseCodeMatch = courseCodePattern.find(fullTitle)
+        val courseCode = courseCodeMatch?.groupValues?.get(1)
+
+        // Extract lecturer information
+        val instructorElement = document.select("td[name=instructorName]").first()
+        val lecturer = instructorElement?.text()?.trim() ?: ""
+
+        // Extract room information
+        val roomElement = document.select("span[name=appoinmentRooms]").first()
+        val room = roomElement?.text()?.trim() ?: ""
+
+        Log.d("DualisService", "Parsed event details: fullTitle='$fullTitle', courseCode='$courseCode', lecturer='$lecturer', room='$room'")
+
+        return EventDetails(
+            fullTitle = fullTitle,
+            courseCode = courseCode,
+            lecturer = lecturer,
+            room = room
+        )
+    }
+
+    /**
+     * Data class to hold detailed event information
+     */
+    private data class EventDetails(
+        val fullTitle: String,
+        val courseCode: String?,
+        val lecturer: String,
+        val room: String
+    )
+
+    /**
+     * Parses concatenated room strings like "HOR-135HOR-136" into separate rooms
+     * separated by commas like "HOR-135, HOR-136"
+     */
+    private fun parseRooms(roomString: String): String {
+        if (roomString.isEmpty()) return roomString
+
+        // Pattern to match room codes like HOR-135, A1.2.03, etc.
+        // This regex looks for patterns that start with letters, optionally followed by digits,
+        // then a dash or dot, then more alphanumeric characters
+        val roomPattern = Regex("([A-Z]+(?:\\d+)?[-.]\\d+(?:\\.\\d+)?)")
+
+        val matches = roomPattern.findAll(roomString)
+        val rooms = matches.map { it.value }.toList()
+
+        return if (rooms.size > 1) {
+            rooms.joinToString(", ")
+        } else {
+            roomString
+        }
+    }
+
+    // Check if the service is properly authenticated
+    fun isAuthenticated(): Boolean {
+        return _authToken != null && !_authToken!!.isEmpty()
+    }
+
+    // Add public method to check demo mode
+    fun isDemoMode(): Boolean {
+        return _isDemoMode
+    }
+
+    private fun isMainPage(html: String): Boolean {
+        val document = Jsoup.parse(html)
+        return document.select("a:contains(Studienleistungen)")
+            .first() != null || document.select("a:contains(Prüfungsergebnisse)")
+            .first() != null || document.select("a:contains(Stundenplan)")
+            .first() != null || document.select("a:contains(Abmelden)").first() != null
+    }
+
+    private fun isTokenInvalidResponse(html: String): Boolean {
+        // Check for common indicators of invalid session/token
+        return html.contains("SESSION_EXPIRED") || html.contains("INVALID_SESSION") || html.contains(
+            "session has expired"
+        ) || html.contains("Sie müssen sich erneut anmelden") || html.contains("Login") && html.contains(
+            "Anmeldung"
+        ) || html.contains("LOGINCHECK") || html.contains("usrname") && html.contains("pass") || html.isEmpty() || html.contains(
+            "error"
+        ) && html.contains("token")
+    }
+
+    private fun reAuthenticateIfNeeded(onComplete: (Boolean) -> Unit) {
+        if (_isReAuthenticating) {
+            // Already re-authenticating, wait for completion
+            return
+        }
+
+        val credentials = _lastLoginCredentials
+        if (credentials == null) {
+            Log.e("DualisService", "No stored credentials for re-authentication")
+            onComplete(false)
+            return
+        }
+
+        _isReAuthenticating = true
+        Log.d("DualisService", "Re-authenticating due to invalid token")
+
+        login(credentials.first, credentials.second) { result ->
+            _isReAuthenticating = false
+            if (result != null) {
+                Log.d("DualisService", "Re-authentication successful")
+                onComplete(true)
+            } else {
+                Log.e("DualisService", "Re-authentication failed")
+                onComplete(false)
+            }
+        }
+    }
+}
