+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
-    <string name="app_name">DHBW Horb</string>
+    <string name="app_name">DHBW Horb Student App</string>
 
     <!-- Login Screen -->
-    <string name="login_title">DHBW Horb Login</string>
+    <string name="login_title">DHBW Horb Dualis Login</string>
     <string name="username">Username</string>
     <string name="password">Password</string>
     <string name="remember_credentials">Remember credentials</string>
